Update README.md - Add DeepLineage plugin
I would like to include the DeepLineage plugin into the readme.md of the Mastodon-Sc organisation to give it some more visibility.

Btw: Is it possible to remove the link http://mastodon.org from the organisation site as? It does not seem to be possible via this readme.md. However, the link currently leads to nowhere, which may be confusing for readers.